requirements.txt
aniso8601==10.0.1
asyncio==3.4.3
babel==2.17.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.0.3
flask-babel==4.0.0
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
pymodbus==3.9.2
pyserial==3.5
pytz==2025.2
PyYAML==6.0.1
six==1.17.0
waitress==3.0.0
Werkzeug==3.1.3
